Added a constructor passing an initializing element to RingBuffer.
In some cases (e.g., shared_ptr) it is not OK to have
un-initialized elements in the RingBuffer.

E.g., when a new shared_ptr is entered into a ring-buffer
slot then it attempts to release the shared_ptr that previously
occupied the slot.

The new constructor lets the user pass an initializer element
which is copied into all slots of the vector.